Using Csampler3f in Shader
