+++ modules/core/Sampler3f.h
@@ -1,6 +1,6 @@
 /**
- * @file CosineSampler.h
- * @brief Cosine Hemisphere  Sampler
+ * @file Sampler3f.h
+ * @brief # dimensional samplers
  * @version 0.1
  * @date 2020-03-11
  * 
@@ -8,12 +8,8 @@
 #pragma once
 #include "types.h"
 #include "random.h"
+#include "Sampler.h"
 namespace rt {
-	struct Sample
-	{
-		Vec3f sample;
-		float pdf;
-	};
 	/*
 	*@brief Returns Samples in 3d from the 2d CSampler Class
 	*/
@@ -33,6 +29,7 @@ namespace rt {
 	class CSampler3fCosine : public CSampler3f
 	{
 	public:
+		//NOTE For CLightSky, use m >> 1 to reduce noise
 		DllExport CSampler3fCosine(float m = 1.0f,std::shared_ptr<CSampler> pSampler = std::make_shared<CSamplerRandom>(4, true))
 			:CSampler3f(pSampler),m(m){}
 		/**
@@ -71,79 +68,26 @@ namespace rt {
 		float m;
 	};
 
-	Vec3f CSampler3fCosine::getNextSample(Vec3f Normal) 
-	{
-		Vec3f k  = getRotationAxis(Normal);
-		Vec2f seed2f = m_pSampler->getNextSample();
-		Vec3f sample3f = conv2cartesian(seed2f[0], seed2f[1]);
-		//return Sample{ rodriguesRotationEqn(k,sample3f), pdf(sample3f) };
-		return rodriguesRotationEqn(k,sample3f);
-	}
-
-	Vec3f CSampler3fCosine::conv2cartesian(float zeta_1, float zeta_2)
-	{
-		/*
-		Samples in sperical coordinates :
-			theta = acos((1 - zeta_1)^(1/(m+1))
-			phi = 2*pi*zeta_2
-		Samples to cartesian coordinates:
-			let A = (1 - zeta_1)^(1/(m+1)
-			x = sin(theta)*cos(phi) = sqrt(1 - A^2)*cos(2*pi*zeta_2)
-			y = sin(theta)*sin(phi) = sqrt(1 - A^2)*sin(2*pi*zeta_2)
-			z = cos(theta)			= A
-
-		*/
-		float A = powf(1 - zeta_1,1/(m+1));
-		float x = cosf(2 * Pif * zeta_2) * sqrt(1 - powf(A,2));
-		float y = sinf(2 * Pif * zeta_2) * sqrt(1 - powf(A,2));
-		float z = A;
-		Vec3f ret(x, y, z);
-		return normalize(ret);
-	}
-
-	float CSampler3fCosine::pdf(Vec3f sample)
-	{
-		float z = abs(sample[2]);
-		float theta = acosf(z);
-		return (1 / Pif) * (z)*sinf(theta);
-	}
-
-	class CSampler3fTangent : public CSampler3f
+		class CSampler3fTangent : public CSampler3f
 	{
 		public :
-			DllExport CSampler3fTangent( std::shared_ptr<CSampler> pSampler = std::make_shared<CSamplerStratified>(4, true, true))
-			:CSampler3f(pSampler){}
+			//NOTE set s = 2 and d = -1 to use with CLightSky
+			DllExport CSampler3fTangent( std::shared_ptr<CSampler> pSampler = std::make_shared<CSamplerStratified>(4, true, true),float s = 1.0f,float d = 0.0f)
+			:CSampler3f(pSampler)
+			,m_s(s)
+			,m_d(d){}
+
+			DllExport CSampler3fTangent( int nSamples ,float s = 1.0f,float d = 0.0f)
+			:CSampler3f(std::make_shared<CSamplerStratified>(nSamples, true, true))
+			,m_s(s)
+			,m_d(d){}
+
 			DllExport virtual Vec3f getNextSample(Vec3f Normal);
 		protected :
 			std::pair<Vec3f, Vec3f> getRandomTangents(Vec3f normal);
+		private :
+			float m_s;
+			float m_d;
 
 	};
-		// TODO: Put in a utilities and merge with the code in CShaderChrome and Shader
-
-	std::pair<Vec3f, Vec3f> CSampler3fTangent::getRandomTangents(Vec3f normal)
-		{
-			std::pair<Vec3f, Vec3f> res;
-			float s1 = random::U<float>();
-			float s2 = random::U<float>();
-
-			if (fabs(normal.val[0]) > 0.1f)
-				res.first = Vec3f(-(s1 * normal.val[1] + s2 * normal.val[2]) / normal.val[0], s1, s2);
-			else if (fabs(normal.val[1]) > 0.1f)
-				res.first = Vec3f(s1, -(s1 * normal.val[0] + s2 * normal.val[2] / normal.val[1]), s2);
-			else
-				res.first = Vec3f(s1, s2, -(s1 * normal.val[0] + s2 * normal.val[1]) / normal.val[2]);
-
-			res.second = normal.cross(res.first);
-			res.first = normalize(res.first);
-			res.second = normalize(res.second);
-			return res;
-		}
-
-	Vec3f CSampler3fTangent::getNextSample(Vec3f Normal)
-	{
-		Vec2f sample = m_pSampler->getNextSample();
-		std::pair<Vec3f, Vec3f> tangents = getRandomTangents(Normal);
-		Vec3f disp = (2 * sample.val[0] - 1) * tangents.first + (2 * sample.val[1] - 1) * tangents.second;
-		return normalize(Normal + disp);
-	}
 }
