+++ modules/core/ShaderChrome.h
@@ -1,36 +1,15 @@
 #pragma once
 
 #include "IShader.h"
-#include "Sampler.h"
+#include "Sampler3f.h"
 
 namespace rt {
-	//namespace {
-		std::pair<Vec3f, Vec3f> getRandomTangents(Vec3f normal)
-		{
-			std::pair<Vec3f, Vec3f> res;
-			float s1 = random::U<float>();
-			float s2 = random::U<float>();
-			
-			if (fabs(normal.val[0]) > 0.1f)
-				res.first = Vec3f(-(s1 * normal.val[1] + s2 * normal.val[2]) / normal.val[0], s1, s2);
-			else if (fabs(normal.val[1]) > 0.1f)
-				res.first = Vec3f(s1, -(s1 * normal.val[0] + s2 * normal.val[2] / normal.val[1]), s2);
-			else
-				res.first = Vec3f(s1, s2, -(s1 * normal.val[0] + s2 * normal.val[1]) / normal.val[2]);
-			
-			res.second = normal.cross(res.first);
-			res.first  = normalize(res.first);
-			res.second = normalize(res.second);
-			return res;
-		}
-	//}
-
 	class CShaderChrome : public IShader
 	{
 	public:
-		DllExport CShaderChrome(CScene& scene, std::shared_ptr<CSampler> pSampler = nullptr)
+		DllExport CShaderChrome(CScene& scene, std::shared_ptr<CSampler3f> dSampler = nullptr)
 			: m_scene(scene)
-			, m_pSampler(pSampler)
+			, m_dSampler(dSampler)
 		{}
 		DllExport virtual ~CShaderChrome(void) = default;
 		
@@ -40,16 +19,13 @@ namespace rt {
 			
 			Vec3f normal = ray.hit->getNormal(ray);								// shading normal
 			if (normal.dot(ray.dir) > 0) normal = -normal;						// turn normal to front
-			const auto tangents = getRandomTangents(normal);					// two random tangent vectors
 			
-			size_t nSamples = m_pSampler ? m_pSampler->getNumSamples() : 1;
+			size_t nSamples = m_dSampler ? m_dSampler->getNumSamples() : 1;
 			for (size_t s = 0; s < nSamples; s++) {
 				
 				Vec3f n = normal;
-				if (m_pSampler) {
-					Vec2f sample = m_pSampler->getNextSample();
-					Vec3f disp = sample.val[0] * tangents.first + sample.val[1] * tangents.second;
-					n = normalize(n + disp);
+				if (m_dSampler) {
+					n = m_dSampler->getNextSample(n);
 				}
 					
 				Ray I = ray.reflected(n);	// reflection vector
@@ -66,7 +42,7 @@ namespace rt {
 		
 	private:
 		CScene& 					m_scene;
-		std::shared_ptr<CSampler>	m_pSampler;
+		std::shared_ptr<CSampler3f>	m_dSampler;
 	};
 }
